Removed full_path from Fileinfo
Removed full_path from the Fileinfo, which was introduced in PHP 8.1 for File Uploads with whole directories. This is not covered in the current Attachament Database schema, which would otherwise lead to an Error. When the directory Upload Feature is needed we have to redesign the Database Structure for it.